Create initial version of footer component; organize styles of main layout components
