+++ src/styles/components/main-layout/header.css
@@ -1,0 +1,11 @@
+.layout-header {
+  width: 100%;
+
+  display: flex;
+  flex-direction: row;
+  gap: 1rem;
+  justify-content: center;
+
+  background-color: var(--layout-bg-color);
+  color: var(--layout-text-color);
+}
